+++ src/main/java/bridge/businesslogic/Bridge.java
@@ -1,6 +1,6 @@
 package bridge.businesslogic;
 
-import bridge.businesslogic.bridgestatus.BridgeCrossingStatus;
+import bridge.businesslogic.bridgehistorystatus.BridgeCrossingHistory;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,14 +8,14 @@ import java.util.List;
 public class Bridge {
     private final List<String> bridge;
     private final List<String> currentlyCrossedBridge;
-    private final BridgeCrossingStatus status;
+    private final BridgeCrossingHistory status;
 
     public Bridge(BridgeMaker bridgeMaker, int size) {
         this.bridge = bridgeMaker.makeBridge(size);
         this.currentlyCrossedBridge = new ArrayList<>();
         StringBuilder upLine = new StringBuilder();
         StringBuilder downLine = new StringBuilder();
-        this.status = new BridgeCrossingStatus(upLine,downLine);
+        this.status = new BridgeCrossingHistory(upLine,downLine);
     }
 
     public boolean isSelectedBridgesRight(String usersPick){
